Reorganize quest subcomponents
