+++ src/components/quests/QuestCosts/index.ts
@@ -1,0 +1,1 @@
+export * from "./QuestCosts";
