+++ src/components/quests/index.ts
@@ -1,3 +1,7 @@
 export * from "./DocumentCard";
 export * from "./ReadingScore";
 export * from "./StatusSelect";
+export * from "./QuestCosts";
+export * from "./QuestTimeRequired";
+export * from "./QuestForms";
+export * from "./QuestUrls";
